# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "description": "This is a sample API documentation",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/books": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Get list of books",
                "description": "Returns list of books",
                "operationId": "GetBookList",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/books-without-borrowed": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Get list of books without borrowed",
                "description": "Returns list of books without borrowed",
                "operationId": "GetBookWithoutBorrowed",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/borrow": {
            "post": {
                "tags": [
                    "Borrow"
                ],
                "summary": "Post Borrowing",
                "description": "Create a Borrowing",
                "operationId": "PostBorrow",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "member_code",
                                    "book_code"
                                ],
                                "properties": {
                                    "member_code": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "M002"
                                    },
                                    "book_code": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "SHR-1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Borrowed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Borrow Book is Successfuly"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/return": {
            "post": {
                "tags": [
                    "Return"
                ],
                "summary": "Post Return a Book",
                "description": "Return a Book",
                "operationId": "PostReturnBook",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "member_code",
                                    "book_code"
                                ],
                                "properties": {
                                    "member_code": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "M002"
                                    },
                                    "book_code": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "SHR-1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Borrowed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Book returned successfuly."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/borrowed-books-count": {
            "get": {
                "tags": [
                    "Members"
                ],
                "summary": "Get List Borrowed Book Count",
                "description": "Returns list  Borrowed Book Count",
                "operationId": "GetBorrowedBookCount",
                "responses": {
                    "200": {
                        "description": "Borrowed Books Count Success"
                    }
                }
            }
        },
        "/api/members": {
            "get": {
                "tags": [
                    "Members"
                ],
                "summary": "Get list of members",
                "description": "Returns list of members",
                "operationId": "GetMemberList",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Books",
            "description": "Books"
        },
        {
            "name": "Borrow",
            "description": "Borrow"
        },
        {
            "name": "Return",
            "description": "Return"
        },
        {
            "name": "Members",
            "description": "Members"
        }
    ]
}